refactor: clean up bindings dir and move tests around
