+++ packages/ai-provider/src/hf-transformers/bindings/registerTasks.ts
