+++ packages/storage/package.json
@@ -5,9 +5,9 @@
   "description": "Ellmers is a tool for building and running DAG pipelines of AI tasks.",
   "scripts": {
     "watch": "concurrently -c 'auto' 'bun:watch-*'",
-    "watch-browser": "bun build --watch --no-clear-screen--target=browser --sourcemap=external --external ellmers-core --outdir ./dist/browser ./src/browser/*/index.ts",
-    "watch-node": "bun build --watch --no-clear-screen--target=node --sourcemap=external --external ellmers-core --outdir ./dist ./src/node/*/index.ts",
-    "watch-bun": "bun build --watch --no-clear-screen--target=bun --sourcemap=external --external ellmers-core --outdir ./dist ./src/bun/*/index.ts",
+    "watch-browser": "bun build --watch --no-clear-screen --target=browser --sourcemap=external --external ellmers-core --outdir ./dist/browser ./src/browser/*/index.ts",
+    "watch-node": "bun build --watch --no-clear-screen --target=node --sourcemap=external --external ellmers-core --outdir ./dist ./src/node/*/index.ts",
+    "watch-bun": "bun build --watch --no-clear-screen --target=bun --sourcemap=external --external ellmers-core --outdir ./dist ./src/bun/*/index.ts",
     "watch-types": "tsc --watch --preserveWatchOutput",
     "build": "bun run build-clean && bun run build-types && bun run build-browser && bun run build-node && bun run build-bun",
     "build-clean": "rm -fr dist/* tsconfig.tsbuildinfo",
